+++ src/components/CardGrid.js
@@ -1,58 +1,21 @@
-import React, {useEffect, useState} from 'react';
+import React from 'react';
 import {
-    DndContext,
-    closestCenter,
-    KeyboardSensor,
-    PointerSensor,
-    useSensor,
-    useSensors, DragOverlay,
-} from '@dnd-kit/core';
-import {
-    arrayMove, rectSortingStrategy,
+     rectSortingStrategy,
     SortableContext,
-    sortableKeyboardCoordinates,
 } from '@dnd-kit/sortable';
 import {SortableItem} from 'components/SortableItem';
 
 import styles from "assets/CardGrid.module.css";
 import {itemsState} from "states";
-import {useRecoilState} from "recoil";
+import {useRecoilValue} from "recoil";
 
 const CardGrid = () => {
-    const [activeId, setActiveId] = useState(null);
-    const [items, setItems] = useRecoilState(itemsState);
-    const sensors = useSensors(
-        useSensor(PointerSensor),
-        useSensor(KeyboardSensor, {
-            coordinateGetter: sortableKeyboardCoordinates
-        })
-    );
-    const handleDragStart = (event) => {
-        setActiveId(event.active.id);
-    };
 
-    const handleDragEnd = (event) => {
-        setActiveId(null);
-        const { active, over } = event;
+    const items = useRecoilValue(itemsState);
 
-        if (active !== over) {
-            setItems((items) => {
-                console.log(active, over);
-                const oldIndex = items.findIndex((item) => item.id === active.id);
-                const newIndex = items.findIndex((item) => item.id === over.id);
-
-                return arrayMove(items, oldIndex, newIndex);
-            });
-        }
-    }
 
     return (
-        <DndContext
-            sensors={sensors}
-            collisionDetection={closestCenter}
-            onDragEnd={handleDragEnd}
-            onDragStart={handleDragStart}
-        >
+
             <div className={styles.container}>
                 <SortableContext
                     items={items}
@@ -72,7 +35,6 @@ const CardGrid = () => {
                 </SortableContext>
             </div>
 
-        </DndContext>
     );
 }
